requirements.txt
SQLAlchemy==1.4.45
uvicorn==0.20.0
fastapi==0.88.0
databases==0.7.0
asyncpg==0.27.0
psycopg2-binary==2.9.5
python-dotenv==0.21.0
alembic==1.9.4
requests==2.28.2
geopy==2.3.0